Support `.scarbignore` files when packaging
This PR makes `scarb package` check for `.scarbignore` files when determining package source files set. This is a trivial addition thanks to features of the `ignore` crate.

Supporting this file is effectively a similar feature to the include/exclude fields of Cargo (https://doc.rust-lang.org/cargo/reference/manifest.html#the-exclude-and-include-fields). This PR fixes #644 then, as we do not want to create multiple ways of achieving the same thing.

commit-id:3e87fb71