Support `.scarbignore` files when packaging (#746)
This PR makes `scarb package` check for `.scarbignore` files when
determining package source files set. This is a trivial addition thanks
to features of the `ignore` crate.

Supporting this file is effectively a similar feature to the
include/exclude fields of Cargo
(https://doc.rust-lang.org/cargo/reference/manifest.html#the-exclude-and-include-fields).
This PR fixes #644 then, as we do not want to create multiple ways of
achieving the same thing.

---

**Stack**:
- #758
- #757
- #748
- #746 ⬅


⚠️ *Part of a stack created by [spr](https://github.com/ejoffe/spr). Do
not merge manually using the UI - doing so may have unexpected results.*