Impl `Clone` for `Filesystem` (#845)
This surprisingly small addition required some bigger changes.
Filesystems are now lifetime-unbound, and rely on internal `Arc`s
instead of references. Technically this means higher heap pressure and
slower code due to increased use of atomics, but `Filesystem`s are by
nature not part of hot paths, so this is fine.

Signed-off-by: Marek Kaput <marek.kaput@swmansion.com>

---

**Stack**:
- #909
- #906
- #892
- #846
- #845 ⬅


⚠️ *Part of a stack created by [spr](https://github.com/ejoffe/spr). Do
not merge manually using the UI - doing so may have unexpected results.*