Remove `update_read_async` method and use `update_read` instead (#924)
The implementation of `update_read_async` caused stack overflows on
Windows only, with exact details being quite a mystery. We did not
really need a super performant async version of this code, as we are
just using it to compute checksums of files residing in well known local
filesystems, and thus the plain old `spawn_blocking` call should do as
fine job as the removed code was expected to.

Signed-off-by: Marek Kaput <marek.kaput@swmansion.com>