Handle case when packaged repo does not have any commits (#956)
Resolves #869.

Unlike cargo which completely ignores VCS when using `--allow-dirty`, we
still want to package the info because that seems reasonably.

There is an edge case when `--allow-dirty` flag which reads a hash of
HEAD commit fails due to repository not having any commits. In that case
we want to treat it like there was no VCS information to include at all
and just ignore it.