Reverse search order in `find_external_subcommand` (#1067)
Although is sounded tempting to allow users to override Scarb extensions
bundled in the default installation, that caused more harm than good in
practice. For example, if the user is working on a custom build of
Scarb, but has another one installed globally (for example via ASDF),
then their custom build would use global extensions instead of the ones
it was built with, which would be very confusing.

---------

Signed-off-by: Marek Kaput <marek.kaput@swmansion.com>