fix: propose consistent usage of debug from tracing (#1324)
Gm guys,

wondering if this is a typo or an actual fix.

In dojo we're using scarb as a dependency, and when we compile a package
of the workspace this error occurs, which prevent us from building a
specific package depending on scarb explicitly.

However, when the whole workspace is compiled, it works fine.

Inspecting the scarb code base, everywhere it seems to be
`tracing::trace`, `tracing::debug` etc... Is this `tracing::log::debug`
forgotten, but not a direct problem as it compiles fine anyway?

Thanks for the feedback!