Update asdf usage instructions on the website/download page (#1960)
If you feel this PR is too minor, on the level of a simple typo fix,
please don’t hesitate to close it.

That said, the `asdf` tool has introduced a few minor changes to its
commands with the version 0.16.0 , and this PR updates the `Install via
asdf` section in the download documentation accordingly. These
adjustments ensure that anyone following the docs will use the correct
commands and avoid potential errors as I did.