Pubgrub lockfile lookup fix (#1957)
fixes #1883

When resolving package dependencies, the resolver accepts workspace
lockfile (Scarb.lock) as one of the inputs. If manifest dependencies can
be locked to concrete revisions defined in the lockfile
(https://github.com/software-mansion/scarb/blob/52c5afe0e26c49e81b93d1b8929dc500a61495e5/scarb/src/core/lockfile.rs#L98)
(i.e. unless the dependency conflicts with the locked version, for
instance because the rev in dependency has been changed by the user
between runs), it should be used by the resolver instead of the most
recent one.

This PR attempts to solve a bug in mechanism of applying locked
revisions in the resolver. When querying the registry for dependency
package summaries
(https://github.com/software-mansion/scarb/blob/47c7c7d8dda31fccc4f63836027977e4ea0f28f4/scarb/src/resolver/algorithm/provider.rs#L181)
we try to lock the manifest dependency to the version specified in the
lockfile. You can see, that we actually request two different
dependencies. If our package A depends on some package B via git
dependency, but B depends on another package C via path dependency, we
attempt to rewrite the dependency between B and C from path to git. This
allows unification if A (or other package in the dep tree) also depends
on C via git dependency. However, we cannot know for sure if this
attempt will be successful, as C may be under B in the directory tree,
thus being shadowed when access via git dependency (we traverse the dir
tree to look for packages, but we do not enter subdirectories of any
directory with Scarb.toml file). Instead, we send to request and use the
rewritten git one if it succeeds, or fallback to path one if not.

This code did not attempt to lock a transitive dependency that has been
rewritten from path to git according to the lockfile, which it should.

Changes:
- **Revert "Temporarily default to primitive resolver (#1892)"**
- **Update error message**
- **Lookup dependency in lockfile after rewriting nested path to git
source when requesting dependency metadata**