Fix dev-deps propagation of compiler plugins for ws members (#2099)
fixes #2118
This problem is analog to the
https://github.com/software-mansion/scarb/issues/1535
(https://github.com/software-mansion/scarb/pull/1542).
If a package (A) that is a ws member depends on another ws member (B),
and (B) depends on some cairo plugin (C) as dev dependency, test
CompilationUnit of (A) will include (C) as dependency of (B), although
dev dependencies of (B) should not be propagated to the test compilation
of (A) (both Cairo and Plugin dependencies).


- **Add failing test for propagation of dev dep plugins from members of
ws**
- **Do not propagate plugins that are dev dependencies of ws members to
packages that depend on these ws members**