feat: doc landing page (#2121)
closes https://github.com/software-mansion/scarb/issues/2094

- Due to formatting restrictions in SUMMARY.md the index page is a
separate file.
- However, creating documentation for each item and creating short
version of the same items is indeed redundant. I did try to use a
HashMap for storing the documentation strings data and retrieve it from
the list if exists insead of generating a new one. After couple of runs
compared time of executions of both versions for Cairo _corelib_ the
highest difference in execution time was around 1 second (6 seconds for
no-hash-map approach, 5 for hash-map approach). I propose the
no-hash-map version as the final solution as the time difference is in
my opinion not significant enough and no extra memory allocation is
needed. Please, let me know if you have different opinion. Keep in mind
that measuring the timing could be slightly different at this point as I
did the tests quite early developing this code.
- If you feel like giving it a try at you local environment be my guest.
If not, you can see demo below.

https://github.com/user-attachments/assets/78d52898-1f3c-485f-bd0e-09b9a9c41962