Disable default features for reqwest again
In #2732 default features were mistakenly re-enabled, which caused `native-tls` crate used by `reqwest` to start looking for `libssl` dynamic libraries on Linux systems. We're using `cross`, which doesn't provide these on its base build images, and we deliberately didn't do anything with this because we decided to use `rustls-tls` instead of `openssl`. That feature modification in turn was done in `scarb/Cargo.toml`. This re-enabling in workspace `Cargo.toml` broke this.