Properly delete Scarb managed cache in tests (#2863)
Temp directory created by `TempDir` is removed from disk when the
associated `TempDir` is dropped.
When running Scarb in tests, new TempDir for Scarb cache is created and
passed via env to the Scarb process.
However, if the TempDir is dropped before calling Scarb, Scarb will
re-create the cache directory on run, never actually deleting it.